pyproject.toml
[build-system]
requires = ["packaging>=24.2", "setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "voidcat"
version = "0.1.0"
description = "A graphical stealth roguelite about a maintenance cat on a dying spaceship."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "OpenAI Codex" }]
dependencies = ["pygame-ce>=2.5"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
]

[project.scripts]
voidcat = "voidcat.gfx_app:main"

[project.optional-dependencies]
dev = ["coverage[toml]>=7.6", "hypothesis>=6.137", "mypy>=1.15", "packaging>=24.2", "ruff>=0.11.0"]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
voidcat = ["assets/gfx/*.json"]

[tool.coverage.run]
branch = true
source = ["src/voidcat"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.mypy]
python_version = "3.10"
files = ["src/voidcat"]
check_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_return_any = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["pygame"]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
